RelayServer: Added CapacityConfig (i.e. maxConnections, maxBytesStored, maxPayloadSize)

+++ opencola-server/relay/common/src/main/kotlin/io/opencola/relay/common/message/v2/store/Usage.kt
@@ -5,4 +5,4 @@ import kotlinx.serialization.Contextual
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Usage(@Contextual val to: Id, val numMessages: Int, val bytesStored: Int)
+data class Usage(@Contextual val to: Id, val numMessages: Int, val bytesStored: Long)
