+++ opencola-server/relay/common/src/main/kotlin/io/opencola/relay/common/policy/Policy.kt
@@ -13,10 +13,10 @@ data class AdminPolicy(
 data class ConnectionPolicy(val canConnect: Boolean = true)
 
 @Serializable
-data class MessagePolicy(val maxPayloadSize: Int = 1024 * 1024 * 50)
+data class MessagePolicy(val maxPayloadSize: Long = 1024 * 1024 * 50)
 
 @Serializable
-data class StoragePolicy(val maxStoredBytes: Int = 1024 * 1024 * 50)
+data class StoragePolicy(val maxStoredBytes: Long = 1024 * 1024 * 50)
 
 @Serializable
 data class Policy(
